ourchive_app/requirements.txt
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
audioread==3.0.0
beautifulsoup4==4.12.2
boto3==1.28.20
botocore==1.31.20
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
cssselect==1.2.0
Django==4.2.4
django-anymail==10.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.2.0
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
EbookLib==0.18
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
exceptiongroup==1.1.1
h11==0.14.0
huey==2.4.5
idna==3.4
iniconfig==2.0.0
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
mobi==0.3.3
mutagen==1.46.0
nh3==0.2.14
openapi-codec==1.3.2
outcome==1.2.0
packaging==23.1
pdfminer.six==20221105
pdfquery==0.4.3
Pillow==10.0.0
pluggy==1.0.0
psycopg2==2.9.7
pycparser==2.21
pydub==0.25.1
pyquery==2.0.0
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
roman==4.1
s3transfer==0.6.1
selenium==4.11.2
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
uritemplate==4.1.1
urllib3>=1.25.4
webdriver-manager==4.0.0
whitenoise==6.5.0
win32-setctime==1.1.0
wsproto==1.2.0
django-apscheduler==0.6.2
unidecode==1.3.6
django-registration==3.4
django-oauth-toolkit>=2.3.0
coverage>=7.4.0